rag/rag-service/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
qdrant-client==1.17.0
llama-index-core==0.12.22
llama-index-embeddings-openai==0.3.1
llama-index-vector-stores-qdrant==0.4.3
openai==1.63.2
xxhash==3.5.0
